Create MatrixColour.py
# Requires the colorama library to work.
# pip install colorama

A comprehensive Matrix Rain effect in python. The program runs over the terminal, and requires monospace fonts to work correctly. Within the file, several modifiers can be updated, like The starting positions, the grid dimensions, raindrop "length", trail colours, etc. 

Use the command line arguments to pass a width, for which it automatically chooses a suitable height for 16:9 screens, and resizes the terminal accordingly to begin.

Debug stats like FPS counter can be enabled by setting the appropriate flags in the program.

## Note- Make sure the terminal window has enough horizontal and vertical spacing to begin. If glitching occurs, resize the terminal contents using (Ctrl + Scroll Wheel up/down) to resize the contents to fit the window, then restart the program.